moved phpdoc2 & phpdoc3 xml conversion to separate classes (XmlToArrayConverters)
