+++ src/phpDocumentor/XmlToArrayConverter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace phpDocumentor;
+
+interface XmlToArrayConverter
+{
+    public function __construct(\SimpleXMLElement $xml);
+
+    public function convert();
+}
